@@ -37,82 +37,85 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.2.12",
-    "@angular/cdk": "^11.2.11",
-    "@angular/common": "^11.2.12",
-    "@angular/compiler": "^11.2.12",
-    "@angular/core": "^11.2.12",
-    "@angular/flex-layout": "^11.0.0-beta.33",
-    "@angular/forms": "^11.2.12",
-    "@angular/platform-browser": "^11.2.12",
-    "@angular/platform-browser-dynamic": "^11.2.12",
-    "@angular/router": "^11.2.12",
-    "@datorama/akita": "^6.1.3",
+    "@angular/animations": "14.2.2",
+    "@angular/cdk": "14.2.2",
+    "@angular/common": "14.2.2",
+    "@angular/compiler": "14.2.2",
+    "@angular/core": "14.2.2",
+    "@angular/flex-layout": "14.0.0-beta.40",
+    "@angular/forms": "14.2.2",
+    "@angular/platform-browser": "14.2.2",
+    "@angular/platform-browser-dynamic": "14.2.2",
+    "@angular/router": "14.2.2",
+    "@datorama/akita": "7.1.1",
     "@electron/remote": "^1.1.0",
-    "@fortawesome/angular-fontawesome": "^0.8.2",
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-regular-svg-icons": "^5.15.3",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
+    "@fortawesome/angular-fontawesome": "^0.11.1",
+    "@fortawesome/fontawesome-svg-core": "^6.2.0",
+    "@fortawesome/free-regular-svg-icons": "^6.2.0",
+    "@fortawesome/free-solid-svg-icons": "^6.2.0",
     "@ngneat/overview": "^2.0.0",
-    "@rx-angular/cdk": "^1.0.0-alpha.8",
-    "@rx-angular/template": "1.0.0-beta.25",
-    "ansi-to-html": "0.6.14",
+    "@rx-angular/cdk": "1.0.0-rc.2",
+    "@rx-angular/template": "1.0.0-rc.2",
+    "ansi-to-html": "0.7.2",
     "collections": "^5.1.12",
     "electron-log": "^4.3.5",
     "electron-store": "^8.0.0",
-    "electron-updater": "^4.3.8",
+    "electron-updater": "^5.2.1",
     "envfile": "^6.14.0",
-    "execa": "^5.0.0",
+    "execa": "^5.1.1",
     "lodash": "^4.17.21",
-    "ng-zorro-antd": "^11.4.1",
-    "ngx-color": "^6.2.0",
-    "rxjs": "~6.6.3",
-    "scheduler": "^0.20.1",
+    "ng-zorro-antd": "14.0.0",
+    "ngx-color": "8.0.3",
+    "rxjs": "~7.5.6",
+    "scheduler": "^0.23.0",
     "subsink": "^1.0.2",
-    "terminate": "2.1.2",
+    "terminate": "2.5.0",
     "tslib": "^2.2.0",
     "uuid": "^8.3.2",
     "wait-on": "^5.3.0",
-    "zone.js": "^0.10.2"
+    "zone.js": "0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.11",
-    "@angular-eslint/eslint-plugin": "4.2.0",
-    "@angular-eslint/eslint-plugin-template": "4.2.0",
-    "@angular-eslint/template-parser": "4.2.0",
-    "@angular/cli": "~11.2.11",
-    "@angular/compiler-cli": "^11.2.12",
-    "@angular/language-service": "^11.2.12",
-    "@nrwl/angular": "12.10.0",
-    "@nrwl/cli": "12.10.0",
-    "@nrwl/cypress": "12.10.0",
-    "@nrwl/eslint-plugin-nx": "12.10.0",
-    "@nrwl/jest": "12.10.0",
-    "@nrwl/linter": "12.10.0",
-    "@nrwl/workspace": "12.10.0",
-    "@octokit/request": "^5.4.15",
-    "@types/jest": "26.0.23",
+    "@angular-devkit/build-angular": "14.2.3",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.2.0",
+    "@angular/compiler-cli": "14.2.2",
+    "@angular/language-service": "14.2.2",
+    "@nrwl/angular": "14.7.5",
+    "@nrwl/cli": "14.7.5",
+    "@nrwl/cypress": "14.7.5",
+    "@nrwl/eslint-plugin-nx": "14.7.5",
+    "@nrwl/jest": "14.7.5",
+    "@nrwl/linter": "14.7.5",
+    "@nrwl/workspace": "14.7.5",
+    "@octokit/request": "^6.2.1",
+    "@types/jest": "28.1.8",
     "@types/lodash": "^4.14.168",
     "@types/node": "14.14.37",
-    "@typescript-eslint/eslint-plugin": "4.22.0",
-    "@typescript-eslint/parser": "4.22.0",
+    "@typescript-eslint/eslint-plugin": "5.37.0",
+    "@typescript-eslint/parser": "5.37.0",
     "conventional-changelog-core": "^4.2.2",
     "copy-webpack-plugin": "6.4.1",
     "cypress": "^6.0.1",
-    "electron": "^12.0.6",
-    "eslint": "7.25.0",
-    "eslint-config-prettier": "8.3.0",
+    "electron": "^20.1.4",
+    "eslint": "8.23.1",
+    "eslint-config-prettier": "8.5.0",
     "eslint-plugin-cypress": "^2.10.3",
     "exitzero": "^1.0.1",
     "generate-package-json-webpack-plugin": "^2.1.2",
-    "jest": "26.6.3",
-    "jest-preset-angular": "8.4.0",
+    "jest": "28.1.3",
+    "jest-environment-jsdom": "28.1.1",
+    "jest-preset-angular": "12.2.2",
     "npm-run-all": "^4.1.5",
-    "nx-electron": "^11.1.1",
-    "prettier": "2.2.1",
+    "nx": "14.7.5",
+    "nx-electron": "^13.2.1",
+    "prettier": "2.7.1",
     "semver": "^7.3.5",
-    "ts-jest": "26.5.5",
-    "ts-node": "~9.1.1",
-    "typescript": "~4.1.5"
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.8.3"
   }
 }
+
